requirements/base.txt
beautifulsoup4==4.3.2
Django==1.5.5
Markdown==2.2.0
Pillow==2.1.0
South==0.8.2
cmsplugin-filer==0.9.5
cmsplugin-news==0.4.2
django-braces==1.2.2
django-classy-tags==0.4
-e git+https://github.com/divio/django-cms.git@3d396c8c1e83478eaf0013f7f2c47925022e1021#egg=django-cms
django-crispy-forms==1.4.0
django-compressor==1.3
django-filer==0.9.5
django-gravatar2==1.1.3
django-timezones==0.2
django-configurations==0.7
dj-database-url==0.2.2

# Strict due to API compatibility issues
django-mptt==0.5.2

django-reversion==1.7.1
django-sekizai==0.7
django-social-auth==0.7.28
django-taggit==0.10
djangocms-admin-style==0.2.0
djangocms-text-ckeditor==2.0.5
djangocms-style==1.0.6
-e git+https://github.com/zerok/django-userprofiles.git@fe29d7feb2a1b0ca8a10be8032f235df5406267f#egg=django-userprofiles
easy-thumbnails==1.4
html5lib==0.99
python-dateutil==2.2
pytz==2013.7
requests==2.0.1

# tablib for export support
tablib==0.9.11

# Search integration
django-haystack==2.1.0

-e git+https://github.com/kliment/pymill.git@f55cab94826168765a5a7d60ffed2641fe89b27d#egg=pymill

# For now we also include Whoosh here. In the future this will
# be moved to dev.txt
Whoosh==2.5.5
